+++ tests/UnitedKingdom/SummerBankHolidayTest.php
@@ -30,7 +30,7 @@ class SummerBankHolidayTest extends UnitedKingdomBaseTestCase implements YasumiT
     /**
      * The year in which the holiday was first established
      */
-    public const ESTABLISHMENT_YEAR = 1965;
+    public const ESTABLISHMENT_YEAR = 1871;
 
     /**
      * Tests the holiday defined in this test.
@@ -39,12 +39,71 @@ class SummerBankHolidayTest extends UnitedKingdomBaseTestCase implements YasumiT
      */
     public function testHoliday()
     {
-        $year = $this->generateRandomYear(self::ESTABLISHMENT_YEAR);
+        $year = $this->generateRandomYear(1970);
         $this->assertHoliday(
             self::REGION,
             self::HOLIDAY,
             $year,
             new DateTime("last monday of august $year", new DateTimeZone(self::TIMEZONE))
+        );
+    }
+
+    /**
+     * Tests the holiday exception in 2020.
+     * @throws \ReflectionException
+     */
+    public function testHolidayBefore1965()
+    {
+        $year = $this->generateRandomYear(self::ESTABLISHMENT_YEAR, 1967);
+        $this->assertHoliday(
+            self::REGION,
+            self::HOLIDAY,
+            $year,
+            new DateTime("first monday of august $year", new DateTimeZone(self::TIMEZONE))
+        );
+    }
+
+    /**
+     * Tests the holiday during trial period in 1965-1970.
+     * @throws \ReflectionException
+     */
+    public function testHolidayTrialPeriod()
+    {
+        $this->assertHoliday(
+            self::REGION,
+            self::HOLIDAY,
+            1965,
+            new DateTime("1965-8-30", new DateTimeZone(self::TIMEZONE))
+        );
+        $this->assertHoliday(
+            self::REGION,
+            self::HOLIDAY,
+            1966,
+            new DateTime("1966-8-29", new DateTimeZone(self::TIMEZONE))
+        );
+        $this->assertHoliday(
+            self::REGION,
+            self::HOLIDAY,
+            1967,
+            new DateTime("1967-8-28", new DateTimeZone(self::TIMEZONE))
+        );
+        $this->assertHoliday(
+            self::REGION,
+            self::HOLIDAY,
+            1968,
+            new DateTime("1968-9-2", new DateTimeZone(self::TIMEZONE))
+        );
+        $this->assertHoliday(
+            self::REGION,
+            self::HOLIDAY,
+            1969,
+            new DateTime("1969-9-1", new DateTimeZone(self::TIMEZONE))
+        );
+        $this->assertHoliday(
+            self::REGION,
+            self::HOLIDAY,
+            1970,
+            new DateTime("1970-8-31", new DateTimeZone(self::TIMEZONE))
         );
     }